Create An Endpoint To Add Inventory To A Project.
The first requirement is to have an API that can add inventory or
'offsets' to a project. This utilizes some parameter conversion plugins
for Roda to sanitize incoming inputs and build a hash structure for a
  new offset object.

Once persisted, the id of the new object is returned along with a links
object to re-request more information about the offset.